requirements.txt
PyQt5>=5.15
bcrypt
pandas
reportlab
matplotlib
pillow
openpyxl
